+++ sources/system_matrix.c
@@ -282,7 +282,7 @@ SystemMatrix add_solid_column
 
     /* TOP */
 
-    if ( ! IS_LAST_LAYER(layer_index, dimensions) )
+    if ( ! IS_LAST_LAYER((unsigned int)layer_index, dimensions) )
     {
         *system_matrix.RowIndices++ = cell_index + layer_offset ;
 
@@ -471,7 +471,7 @@ SystemMatrix add_liquid_column_mc_4rm
 
     /* TOP */
 
-    if ( ! IS_LAST_LAYER(layer_index, dimensions) )
+    if ( ! IS_LAST_LAYER((unsigned int)layer_index, dimensions) )
     {
         *system_matrix.RowIndices++ = cell_index + layer_offset ;
 
@@ -615,7 +615,7 @@ SystemMatrix add_liquid_column_mc_2rm
 
     /* TOP */
 
-    if ( ! IS_LAST_LAYER(layer_index, dimensions) )
+    if ( ! IS_LAST_LAYER((unsigned int)layer_index, dimensions) )
     {
         *system_matrix.RowIndices++ = cell_index + 2 * layer_offset ;
 
@@ -724,7 +724,7 @@ SystemMatrix add_bottom_wall_column_mc_2rm
 
     /* Top connected to Virtual Wall */
 
-    if ( ! IS_LAST_LAYER(layer_index, dimensions) )
+    if ( ! IS_LAST_LAYER((unsigned int)layer_index, dimensions) )
     {
         *system_matrix.RowIndices++ = cell_index + layer_offset ;
 
@@ -744,7 +744,7 @@ SystemMatrix add_bottom_wall_column_mc_2rm
 
     /* Top connected to Channel */
 
-    if ( ! IS_LAST_LAYER(layer_index, dimensions) )
+    if ( ! IS_LAST_LAYER((unsigned int)layer_index, dimensions) )
     {
         *system_matrix.RowIndices++ = cell_index + 2 * layer_offset ;
 
@@ -869,7 +869,7 @@ SystemMatrix add_top_wall_column_mc_2rm
 
     /* Top connected to Silicon */
 
-    if ( ! IS_LAST_LAYER(layer_index, dimensions) )
+    if ( ! IS_LAST_LAYER((unsigned int)layer_index, dimensions) )
     {
         *system_matrix.RowIndices++ = cell_index + layer_offset ;
 
@@ -1023,7 +1023,7 @@ SystemMatrix add_virtual_wall_column_mc_2rm
 
     /* TOP */
 
-    if ( ! IS_LAST_LAYER(layer_index, dimensions) )
+    if ( ! IS_LAST_LAYER((unsigned int)layer_index, dimensions) )
     {
         *system_matrix.RowIndices++ = cell_index + layer_offset ;
 
@@ -1162,7 +1162,7 @@ SystemMatrix add_liquid_column_pf
 
     /* TOP */
 
-    if ( ! IS_LAST_LAYER(layer_index, dimensions) )
+    if ( ! IS_LAST_LAYER((unsigned int)layer_index, dimensions) )
     {
         *system_matrix.RowIndices++ = cell_index + layer_offset ;
 
@@ -1272,7 +1272,7 @@ SystemMatrix add_bottom_wall_column_pf
 
     /* Top connected to Virtual Wall */
 
-    if ( ! IS_LAST_LAYER(layer_index, dimensions) )
+    if ( ! IS_LAST_LAYER((unsigned int)layer_index, dimensions) )
     {
         *system_matrix.RowIndices++ = cell_index + layer_offset ;
 
@@ -1292,7 +1292,7 @@ SystemMatrix add_bottom_wall_column_pf
 
     /* Top connected to Channel */
 
-    if ( ! IS_LAST_LAYER(layer_index, dimensions) )
+    if ( ! IS_LAST_LAYER((unsigned int)layer_index, dimensions) )
     {
         *system_matrix.RowIndices++ = cell_index + 2 * layer_offset ;
 
@@ -1417,7 +1417,7 @@ SystemMatrix add_top_wall_column_pf
 
     /* Top connected to Silicon */
 
-    if ( ! IS_LAST_LAYER(layer_index, dimensions) )
+    if ( ! IS_LAST_LAYER((unsigned int)layer_index, dimensions) )
     {
         *system_matrix.RowIndices++ = cell_index + layer_offset ;
 
@@ -1522,7 +1522,7 @@ SystemMatrix add_virtual_wall_column_pf
 
     /* TOP */
 
-    if ( ! IS_LAST_LAYER(layer_index, dimensions) )
+    if ( ! IS_LAST_LAYER((unsigned int)layer_index, dimensions) )
     {
         *system_matrix.RowIndices++ = cell_index + 2 * layer_offset ;
